Move runtime scanner code directly into scanner module
