+++ src/core/scan/alphabet.rs
